Change `Bugsnag.TaskSupervisor` restart strategy to `:temporary`
This will stop the infinite retry loop of reports if for some reason bugsnag is
unavailable. Without that, it is possible to *crash* our app if, for some
reason, reports are imposible.

This is in part mitigated by the fact that `Task`s stop restarting after their
timeout, which default to 5s. That means, we'd need to accumulate enough error
reports in 5 seconds to have their restarts crashing the `TaskSupervisor`.

That would be indeed a terrible place to be.

Signed-off-by: Milhouse <renanranelli@gmail.com>